Add loader to borrow button Sorted collaterals list by amount in wallet Move summary to separate file
