+++ src/pages/tokenLending/BorrowTokenPage/InstantBorrowContent/InstantBorrowContent.tsx
@@ -1,16 +1,10 @@
-import { FC } from 'react'
-
 import { useWallet } from '@solana/wallet-adapter-react'
 
 import { Button } from '@banx/components/Buttons'
-import { StatInfo, VALUES_TYPES } from '@banx/components/StatInfo'
-import { DisplayValue } from '@banx/components/TableComponents'
-
-import { BorrowSplTokenOffers, CollateralToken } from '@banx/api/tokens'
 
 import { Separator } from '../components'
 import InputTokenSelect from '../components/InputTokenSelect'
-import { getSummaryInfo } from './helpers'
+import { Summary } from './Summary'
 import { useBorrowTokensList, useCollateralsList } from './hooks/useCollateralsList'
 import { useInstantBorrowContent } from './hooks/useInstantBorrowContent'
 
@@ -35,7 +29,8 @@ const InstantBorrowContent = () => {
     borrowTokenBalanceStr,
 
     errorMessage,
-    executeBorrow,
+    borrow,
+    isBorrowing,
   } = useInstantBorrowContent()
 
   const { collateralsList } = useCollateralsList()
@@ -71,9 +66,10 @@ const InstantBorrowContent = () => {
 
       <Summary collateralToken={collateralToken} offers={offers} />
       <Button
-        onClick={executeBorrow}
+        onClick={borrow}
         disabled={!wallet.connected || !!errorMessage}
         className={styles.borrowButton}
+        loading={isBorrowing}
       >
         {!wallet.connected ? 'Connect wallet' : errorMessage || 'Borrow'}
       </Button>
@@ -82,42 +78,3 @@ const InstantBorrowContent = () => {
 }
 
 export default InstantBorrowContent
-
-interface SummaryProps {
-  offers: BorrowSplTokenOffers[]
-  collateralToken: CollateralToken
-}
-
-export const Summary: FC<SummaryProps> = ({ offers, collateralToken }) => {
-  const { upfrontFee, weightedApr, weeklyFee } = getSummaryInfo(offers, collateralToken)
-
-  const statClassNames = {
-    value: styles.fixedStatValue,
-  }
-
-  return (
-    <div className={styles.summary}>
-      <StatInfo
-        label="Upfront fee"
-        value={<DisplayValue value={upfrontFee} />}
-        tooltipText="1% upfront fee charged on the loan principal amount, paid when loan is funded"
-        classNamesProps={statClassNames}
-        flexType="row"
-      />
-      <StatInfo
-        label="Est weekly fee"
-        value={<DisplayValue value={weeklyFee} />}
-        tooltipText="Expected weekly interest on your loans. Interest is added to your debt balance"
-        classNamesProps={statClassNames}
-        flexType="row"
-      />
-      <StatInfo
-        label="Weighted APR"
-        value={weightedApr / 100}
-        valueType={VALUES_TYPES.PERCENT}
-        classNamesProps={statClassNames}
-        flexType="row"
-      />
-    </div>
-  )
-}
